access

+++ library_project/library_app/views.py
@@ -2,7 +2,8 @@ from django.shortcuts import render, get_object_or_404, redirect
 from .models import Book, IssuedBook
 from .forms import BookForm
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
 # Create your views here.
 
 
@@ -18,6 +19,7 @@ def profile(request):
     title = "Profile"
     return render(request, 'library_app/profile.html', {'title': title})
 
+@login_required
 def book_list(request):
     books = Book.objects.all().order_by('title')
     paginator = Paginator(books, 4)  # Show 10 books per page
@@ -32,6 +34,7 @@ def book_list(request):
     title = "Book List"
     return render(request, 'library_app/book_list.html', {'books': paginated_books, 'title': title})
 
+@login_required
 def book_detail(request, id):
     book = get_object_or_404(Book, id=id)
     title = "Book Detail"
@@ -41,7 +44,10 @@ def issued_books(request):
     title = "Issued Books"
     return render(request, 'library_app/issued_books.html', {'title': title})
 
+
 def book_create(request):
+    if not request.user.is_staff:
+        return HttpResponseForbidden("You are not authorized to add books.")
     if request.method == 'POST':
         form = BookForm(request.POST)
         if form.is_valid():
@@ -54,7 +60,10 @@ def book_create(request):
 
     return render(request, 'library_app/book_create.html', {'form': form, 'title': title})
 
+
 def book_update(request, id):
+    if not request.user.is_staff:
+        return HttpResponseForbidden("You are not authorized to add books.")
     book = get_object_or_404(Book, id=id)
     if request.method == 'POST':
         form = BookForm(request.POST, instance=book)
@@ -67,7 +76,10 @@ def book_update(request, id):
     title = "Update Book"
     return render(request, 'library_app/book_update.html', {'form': form, 'title': title})
 
+
 def book_delete(request, id):
+    if not request.user.is_staff:
+        return HttpResponseForbidden("You are not authorized to add books.")
     book = get_object_or_404(Book, id=id)
     if request.method == 'POST':
         book.delete()
